Add a client-side palette editor
This makes it possible to add and edit palette entries. Unfortunately,
since our palette elements  are still sort of divided across
client/server, this doesn't yet  allow you to add a new underlying IR
model (the server endpoint for that  exists, but I want to make sure
this is really the right approach.