Progress towards plotting a sum of models
For the demo, we will want to be able to plot a _sum_ of data traces.
The two challenges here are

  * how to have a good UI interface
  * how to deal with summing data traces with different domains

This commits tackles a bit of both. For the we switch to using a table
based format (where the data trace cell will eventually support multiple
lines of summed data traces). For summing data traces, I started by
adding functionality for the simplest thing possible: sum up series by
taking the union of their domains, doing linear interpolation on missing
points.