Patch cookie-related race condition
This doesn't outright fix all issues, but it definitely makes it a lot
harder to hit the problem in the first place. The bad behaviour is when
multiple requests are made without a cookie. Then, there is a race to
set a cookie, and we end up creating several cookies

We do two things to fix this:

   - don't make a cookie for serving up static content (where racing is
     a desired feature)

   - overwrite cookies (by explicitly setting the path) so that only one
     cookie wins in the end